third_party/blink/web_tests/virtual/css-sizing-keyword-animation/external/wpt/css/css-sizing/animation/max-height-interpolation-expected.txt

This is a testharness.js-based test.
Found 126 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 26px + size * - 0.3 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 20px + size * 0 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( min - content , 14px + size * 0.3 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (0.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 10px + size * 0.5 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (0.6) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 8px + size * 0.6 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (1) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] Web Animations: property <max-height> from [20px] to [min-content] at (1.5) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 10px + size * 1.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [min-content] to neutral at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 3px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [min-content] to neutral at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [min-content] to neutral at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 3px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [min-content] to neutral at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 3px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [min-content] to neutral at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [min-content] to neutral at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 3px + size * 0.7 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 3px + 0px + size * 1.3 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 3px + 0px + size * 0.7 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , - 3px + 0px + size * 1.3 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (0) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (0.3) should be [min-content]
  assert_equals: expected "min - content " but got "calc - size ( min - content , 3px + 0px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (-0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 13px + size * - 0.3 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (0) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 10px + size * 0 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 7px + size * 0.3 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (-0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 13px + size * - 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 10px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 7px + size * 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + 0px + size * 0.5 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + 0px + size * 0.6 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + 0px + size * 1.5 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + 0px + size * 0.5 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + 0px + size * 0.6 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + 0px + size * 1.5 ) "
Harness: the test ran to completion.
